backend: add the date to the .at command
This is a rework of the .at command's regex to match a reminder time. It
adds the possibility to use either a date only, a date with a time, or a
time with a date.

This makes the .at command more flexible and more on-part with Sopel
7.1's current .at command.

Thanks to dgw for the reviews!

Co-authored-by: dgw <dgw@technobabbl.es>