plugin, test_integration: Update to use Sopel 8 connection check style
As of Sopel 8.0, `bot.connection_registered` is a read-only property
whose value depends on two underlying booleans: one in the IRC backend
(connection state) and one in the bot itself (registration state).

That semantic change means depending on `sopel>=8.0` is *really*
necessary now (not just for tests, but at runtime).

----

Rant time:

I *do not like* messing with the internal attributes (we should also
make `backend.connected` a property, really) but mocking a read-only
property is surprisingly obnoxious.

`unittest.mock.patch` and `pytest`'s `monkeypatch` fixture both seem to
be incapable of doing this. It *should* be simple, but nope. I spent
over an hour reading about & trying those.

It's possibly doable with `unittest.mock.PropertyMock`, but (for now, at
least) I simply couldn't figure out how to combine that with the
`botfactory` already used here.