meta: upgrade pyproject.toml to modern standards
This follows the modernization of the pyproject.toml file of Sopel
itself:

* using dependency groups (with the appropriate Py3.8 shim)
* moving all configuration to pyproject.toml and removing setup.cfg
* adapting the CI and the Makefile

Thanks to all the folk who contributed to that modernization of Sopel,
as I merely copied their work into this project.

Note about license: pip's latest version supporting Py3.8 needs the
old-style declaration of license (with a text attribute). So until Py3.8
is dropped, it'll be like that. This may become a problem at some point,
something that is discussed in Sopel's issue 2696.

Issue 2696: https://github.com/sopel-irc/sopel/issues/2696